+++ src/components/App/Navbar/MenuIcon/index.js
