+++ src/components/App/Navbar/index.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import style from './Navbar.module.css'
 import Form from "../Form";
-import MenuIcon from "../MenuIcon";
-import LogoutButton from "../LogoutBtn";
+import MenuIcon from "./MenuIcon";
+import LogoutButton from "./LogoutBtn";
 
 
 function Navbar({opened, setOpened}) {
